requirements.txt
aiosmtplib==2.0.2
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
black==24.8.0
blinker==1.8.2
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-mail==1.4.1
filelock==3.15.4
flake8==7.1.1
greenlet==3.0.3
h11==0.14.0
identify==2.6.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.8.0
prometheus_client==0.20.0
pyasn1==0.6.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyflakes==3.2.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.32
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.5
validate-docbr==1.10.0
virtualenv==20.26.3
WTForms==3.1.2
